Refactor : "Add homepage" add Homepage and refactoring move MyHomePage Class from main to home
